pyproject.toml
[tool.poetry]
name = "ai-assistant"
version = "0.1.0"
description = "AI assistant service using SmolagentS and FastAPI"
authors = [
    "Luis Otavio <luisotaviomgr1@gmail.com>"
]
readme = "README.md"
packages = [{include = "ai_assistant", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = "^0.115.10"
uvicorn = "^0.34.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
smolagents = {extras = ["litellm"], version = "^1.9.2"}
sqlalchemy = "^2.0.38"
pydantic-settings = "^2.8.1"
python-multipart = "^0.0.20"
python-magic = "^0.4.27"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "ai_assistant.main:main"